Remove cfg_if from src/unix/notbsd/android/
